chatbot/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.2
pyyaml==6.0.1
markdown==3.5.1
gunicorn==21.2.0
